requirements.txt
accelerate==0.26.1
albumentations==2.0.3
bitsandbytes==0.45.2
braceexpand==0.1.7
decord==0.6.0
diffusers==0.29.1
dnnlib==0.0.1
einops==0.8.1
huggingface_hub==0.24.5
imageio==2.33.1
numpy==2.2.2
omegaconf==2.3.0
open_clip_torch==2.24.0
opencv_contrib_python==4.9.0.80
opencv_contrib_python_headless==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.11.0.86
packaging==24.2
peft==0.12.0
Pillow==10.1.0
Pillow==11.1.0
safetensors==0.5.2
setuptools==60.2.0
tabulate==0.9.0
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.2
transformers==4.44.0
webdataset==0.2.100
